First commit as new project. Changed references to chyves.
Changed command from `iohyve` to `chyves`. Named after the plant,
similar to the onion. Spelt incorrectly to mimic the bhyve project.
Originally was going to use iochyve but decided that it would easily
get confused. A memorable logo is in the works.

Created new document structure and folder. Root folder contains
documents about the project and the Makefile. The installed files are
now in their respective places `sbin/` and `man/`

Changes other internal references to `iohyve` to `chyves`. For example,
the VMM names now reference `chy-$guest`. Still need to rebuild man
page.