Fix: Select correct guest when grep-ing data.
This resolves an issue where the wrong data (parameters & disk image)
is used when a duplicate $name dataset exists on the system. This patch
does this by grep-ing for chyves/guests/$name rather than $name. This
was observed to happen
when the same dataset as $name exists on system (2 or more pool) and
when the non-intended pool is alphabetically before the intended pool.
This was first observed when starting and setting
properties.

For a demonstrated example of problem:

root@bhost:~ # zfs list -o name -r | grep bguestname
apool/backuphyve/guests/bguestname
apool/backuphyve/guests/bguestname/disk0
zpool/chyves/guests/bguestname
zpool/chyves/guests/bguestname/disk0
root@bhost:~ # iohyve start bguestname

In the example above (before this patch), bguestname would start using
the parameters from dataset backuphyve/guests/bguestname
and the disk image from backuphyve/guests/bguestname/disk0. However,
chyves list shows only datasets grep-ed with chyves/guests/$name.

It makes sense to use chyves/guests/$name for active guests so that a
replicated copy can exist on the same system without interference. This
should not interfere with multiple pool configurations.

This was first reported on April 5th, 2016 but as of the creation of
this project the PR #127 was not pulled in. This will cause an issue
for chyves guests if this PR is not pulled in with iohyve and share
similar names.

This commit also begins to add using a new dataset $pool/chyves/guests.