Enh: Rewrite of info() to use flags for verbosity
Complete rewrite of chves info to use flags to
change the verbosity level of the output. Uses
flags -v -s -d and -l which must be used as one
series of characters after the "-". So "-dsl", "-v"
"-dvpls" all work, "-d -v -s -l" will not work.
Order of flags is not important.

 -v is for verbose which adds the fields: OS,
    Loader, Tap, and Con.

 -s is for status which add the fields: VMM,
    Running, and rcboot.

 -d is for description

 -l is for less and pipes the final output to less

The way this is written, extra characters are
ignored.

Docs also rewritten in this commit. Man page needs rebuilding.