Enh: Created .info dataset, moved .defaults
Created .info dataset to keep important properties which currently
includes chyves:dataset_version and chyves:dataset_role. Easy to grep
and inverse grep.

chyves:dataset_version will contain the version the dataset. This will
be set upon dataset creation and then updated via chyves-utils using
the “chyves-upgrade” command. This will be the automated process of
ensuring the dataset contains the necessary properties to run
correctly. This will also make future dataset changes easier to
implement.  There is also a check ran before anything else to ensure
corruption does not occur. Setup is still able to run if multiple
datasets exist. The setup use case is if you want to migrate guests
from one pool to a different one that is not currently setup, limited
but possible.

chyves:dataset_role will contain which role the dataset is used for.
The valid values are “primary”, “secondary”, and possibly “offline” in
the future. The primary pool will always contain the ISO and Firmware
resources. All pool that are active and not the primary are considered
“secondary”.

.defaults dataset was moved as a child of guests/ because it really
makes more sense as the defaults are for guest properties. This negates
the need to making any changes to utilizing the “chyves set” command.
Only the primary pool will have the .defaults dataset as a single point
of management.

Also added version variable needed for various parts of above
functionality and changed __version() to manage from a single block of
code.