ENH: Added func to list pools
Added function __get_pool_names to list all pool name. The “-a” flag
displays all pools even offline pools. The “-i” flag only displays
pools in an offline state. Not using a flag will display pools in the
primary or secondary state. The primary|secondary|offline state refers
to the chyves property “dataset_role” not to the ZFS state of a pool.