ENH: Func to get mount path for guest
This function __get_path_for_guest_dataset returns the mount path
location of a guest. This is used in the future extensively in the
__load function to reduce code. There are also several places where
this can be used to properly use the correct dataset when saving files.
This is currently mishandled and everything is saved on the primary
pool rather than the proper guest. The could cause problems when moving
a guest and losing grub.cfg/device.map files.