FIX: Wrote func to run all kernel modules check
Did not test the application of __verify_kernel_module_loaded beyond
the debug module. Tried starting a guest today and realized the error
of my ways. Wrote __verify_all_kernel_modules for proper application
of what I was trying to achieve. The __verify_kernel_module_loaded
function can still be called if a module needs to be checked, loaded
or unloaded.

Getting a lot of output when starting a guest because the above
function gets called three times before starting. Still thinking of
a soltuion.

Various white spacing issues.