DOC: Normalized white-spacing.
Normalized white-spacing for files. Most of the edits I have done have
had a mix of tabs and spaces. This is due to my inattentiveness while
using Atom. This commit is to resolve that.

Everything has been normalized to use spaces for text files such as the
man page and *.md files. This is important when output of the text is
static and appearance is important.

Scripts will use tabs. This makes it easier to read, edit, and develop
from the command line in an editor. There should never be the internal
conversation "Is it two or three spaces after an `if` statement?". It
is a tab, every time.

man page now uses spaces consistently.

Added section to DEVELOPMENT.md to address this.

Rebuild man page.

Updated CHANGELOG.md to reflect these changes.

Start to use longer version integer number so it is easier to compare.
Need to include this as a property.