FIX: Various small fixes with funcs
Fixed __multi_chyves_zfs_property as the set value was not configured.

Repaired __list to work correctly. Rearranged flags to that the -a flag
better indicates that it only controls the fields that show, not the
guest list. The -k -i and -g flags control which guests and disk are
shown.

Added -n flag for notes property.