ENH: Added ability to have boot priority.
The chyves property boot now influences the boot order of the guests.
With “0” being do not boot and any positive integer will start the
guest during the host’s boot process. The order is determined by the
guests with the highest priority set first then down to the last guests
with the lowest positive integer. If two or more guests share the same
boot priority value then the guests are started in alphabetical order.
Guests in offline or template mode will not start due to the main
chyves application catching it and giving an error.

A sleep value of 8 was given for the networking changes to take and 3
second between guest start. These are precautions and perhaps not
necessary.

An explanation on how the $guestlist variable is set:

# Get the ZFS dataset name and the value (-o name,value) for
chyves:boot for every filesystem (-t filesystem) and put it in
scriptable form (-H).
zfs get -H -t filesystem -o name,value chyves:boot

# Grep for all chyves/guests/, by using the trailing
“/“ the parent datasets are not displayed.
grep chyves/guests/

# Sort column two (-k2) in numeric (n) reverse (r) mode then sort
column two as text (-k1)
sort -k2nr -k1

# If the second column is greater than 0 then print the dataset name
from column 1.
awk '{ if ($2 > 0) print $1; }'