ENH: Added developer mode
Added developer mode, this is determined by the value set for
“dev_mode” on “.config” of the primary pool. Must be “on” to activate
features.

Features implemented with this commit:

Display the full bhyve command used to start the guests just before
executing the command. This is done in __boot, __start_ and __uefi.

Allow functions to be called straight from the command line with
“chyves dev”. For example: “chyves dev __version” executes the
__version() function. Additional parameters are supported as well due
to the function indexed switch over. So this works too: “chyves dev
__list tap active”.