ENH: New properties for network expansion.
These are the properties that are going to be used for the new network
enhancements.

In the future, the tap’s membership is declared, bridges will have the
option of which (if any) physical members are assigned to it. There is
the possibility of using VALE switching but I am unsure how that is
implemented.

Line 794 is necessary as _PRIMARY_POOL will not be set if running
chyves setup for the first time.