ENH: Partial rewrite of __setup()
Added use of __multi_chyves_zfs_property, this drastically increases
readability of properties and their values.

Uses better way to eval if primary pool is set.

With recent changes to __verify_valid_pool, running setup pool on a
already setup pool does not clobber the pool and does not set a primary
pool as secondary.