ENH: Added multi-guest support to __create
This allows for multiple guests be specified in the command using
commas to separate each guest. If  you want to create 3 guests with 8G
hard drives on pool “ssd-pool”, the command would be:

    chyves create g1,g2,g3 8G ssd-pool

Each guest name is verified to not be used before trying to create it.

Added more verbosity to let the end user know what is going on.