ENH: Added multi-guest support to __cloneguest
This allows for multiple guests be specified in the command using
commas to separate each guest. If  you want to clone 3 guests from
goldenVM to pool “dev-pool”, the command would be:

chyves clone goldenVM devteam1,devteam2,devteam3 dev-pool

Each guest name is verified to not be in use before trying to create it.

A new UUID is used for each guest created using the -r flag. Using the
-c flag it is an exact clone as of the moment the snapshot was taken.
The -c method is not recommended unless you know what you are doing.

$_UUID_GENERAL_USE is used as part of the snapshot name because of the
statically improbably  chance of a collision. The clone name was used
before but this could lead to problems if you clone the guest
frequently to the same clone name if the zfs destroy command failed.

Added more verbosity to let the end user know what is going on.