ENH: Added __verify_iommu_capable func
Added __verify_iommu_capable to verify machine has I/O MMU aka VT-d aka
AMD-Vi support. This will be only be used when PCI passthrough is used
by the guest.

Thanks to @churchers for pointing this detection method.

This is in reference to this issue raised on pr1ntf/iohyve:
https://github.com/pr1ntf/iohyve/issues/166