Deprecated "persist" property
Deprecated the “persist” property because it causes more trouble than
it is worth. This is due to bhyve exiting with a “0” for reboots, this
is great and seems to be a unique code. However bhyve will exit with a
“1” for many reasons. If bhyve would use a unique exit code it would
make this property and functionality viable. Otherwise you are just
asking from problems because a “missing data backing” error also exits
with a code “1”, an infinite loop of that guest trying to boot with no
dataset is not acceptable and not what a reasonable person would
expect. In the future __start() function, when an exit code “0” from
bhyve is detected: the loader is reloaded and bhyve is reran otherwise
the guest’s VMM is destroyed. This is normal behavior of a VM on other
platforms, persist functionality is atypical.