DOC: New __start() is now written.
New __start() has been tested fairly well but I need to do a heavy
shake down of it before releasing.

Works much better than before and reboots guest correctly and VMM
resources are destroyed when bhyve exits due to anything but a reboot.
More in the actual commit of the code.