ENH: Added func __generate_bhyve_slot_string
Added function __generate_bhyve_slot_string to generate the bhyve
command string when supplied with a list of PCI devices.

This bit of clever code will generate and return the output as a global
variable. Optimized for best guest performance in terms of using PCI
slots and when needed PCI functions. Also handles custom PCI dev fairly
well.

This expands a great deal on the simple __get_bhyve_cmd() function.