ENH: Added func __generate_bhyve_disk_string
Generates a string for each disk device.

UEFI guests are limited to using slots 3-7 and slot 3 is reserved (in
chyves) to be used by the optical device. Special handling is in place.

By default the ZFS volblocksize value is pulled from each disk and
passed to bhyve as the sector size for performance reasons.