ENH: Commit for new network handling methods.
This code has been sitting for a few weeks now uncommitted. There are
probably a few bugs remaining, but for the most part most have been
ironed out.

This massive commit encompasses the new handling methods used for guest
networking. This is a massive expansion from the previously handling
method.

Summary: VALE support added. Command interface for handling the network
design. Network design is recreated and verified when starting guest.
Private bridges.

There are bits of clever code here but it is overwhelming how much code
was written for something that was previously simple.