Deprecated __rmpci
No function to replace it but is addressed in man page under Manually
Setting Properties section.

Do not see why I would keep this around when use case is very narrow
and is not a complete suite of functionality. Like I should add the
ability to add PCI dev as well.