ENH: New property chyves_guest_version
This encompasses all changes for this feature.

This is needed because in the future moving guests between other chyves
installations needs to be possible and a mechanism needs to be able to
deal with version differences so that chyves does not reference
properties that do not exists, have changes, or that have a new format
under the same property name.

Limits guests from starting if on an older major version. Warns if on
an older minor version or attempting to run a new version.

Upgrade method similar to dataset upgrade method (future).

Added to verbose flag ‘-v’ in ‘chyves info’.

Man page changes and __help().

Added to guest properties for ‘chyves <guest> create’ and adjusted
whitespace.

Added user input handling.

Added ‘chyves <guest> upgrade’ call in __parse_cmd_ingress(), kind of
hacky due to the way git diffs work. Will re-add reclaim shortly.