ENH: Added 'chyves dataset <pool> upgrade'
This will incrementally upgrade a dataset and then the guests to the
latest dataset_version and chyves_guest_version respectively. This is
an upgrade mechanism to ensure the correct parameters are available.

Guests are only upgraded when the primary pool is on the latest
version. There is a possibility of running the upgrade on a secondary
pool first, which when then set new guest parameters with a null value
due the primary pool not actually contain new defaults parameters. This
is the reason for lines 181 and 182.

Also refined dataset check in __preflight_check(). More verbose error
message, less to check in ‘if’ statement. Also can run ‘chyles list
pools’ to check which pools do not have the latest version.