FIX: Do not let newer chyves_guest_version guests start.
This is due to the possibility of making a large change in the future
and not being able to start a guest. Or possibly switching key value
storage mechanisms to something like UCL.

For example: chyves version 999.99.999 on new UCL version 9999 would
attempt to start on this chyves version 0.0.11 and cause issues.

Unfortunately previous versions will have issues and fail without
designed warning messages. Guess it is good thing this is pre-public
release.