ENH: Added limited support for raw file backed disks for guests.
ALL FILES in the folder ‘/chyves/<pool>/guests/<guest>/img’ and child
folders are attached to the guest when started. Manually adding a
description or note annotation is possible for raw image files that do
not contain a period ‘.’. The Bourne Shell will not export a variable
with a period in the name.

Renamed and reworked __gvset_guest_disk_list(), formerly
__return_guest_disk_list(). Now includes raw file disk images and is
utilized more in script.

Reworked __gvset_first_disk() to use __gvset_guest_disk_list().

Created __gvset_disk_clone_origin_for_info_func() to find origin of
disk. Used for both the guest and each disk.