ENH: Added feature to revert to snapshot on (re)boot.
With this feature a guest is reverted to a previous snapshot state. The
guest can be reverted on boot, reboot, both, or the feature can be
turned off. A specific snapshot can be specified if automated snapshots
are taken globally on the machine or the ‘revert_to_snapshot’ property
can be left empty which then the last snapshot is the reversion point.
This completely reverts that guest to the state including properties.
This is helpful for guests that get modified outside of specification
such as frequently compromised web server or VDI instances.

This commit increments the ‘chyves_guest_version’ and ‘dataset_version’
versions. To handle the new properties and also include the ‘-S’ flag
in ‘bargs’.