FIX: Reverting snapshot would not reload guest settings.
This fixes an issue when reverting a guest on start or reboot where the
guest properties would not be reloaded. This is critical because a
guest could be reverted to a state with a different OS (extreme) or
less extreme, the CPU count changed. Without this fix an observed
problem would be hard to diagnose as the problem might only manifest on
the first reversion until the guest was stopped and started again.

Added notation to indicate which reversion is happening.

Moved guest running check to before the reversion as reverting a live
guest might be too interesting.