FIX: Issue when reverting guest with different network config. CGV+
Increments chyves guest version to 0200.

If a guest had a different network configuration at a snapshot state,
then a discrepancy is created with the network associations when
reverted. This fix attempts to unbreak the associations, however it
does not revert the guest to the exact network configuration on the
host as that is not tracked.

The old configuration is pulled from the .zfs folder at the base of the
dataset which contains the files at different snapshot states. This
commit turns the ‘snapdir’ zfs property ‘on’ and increments the chyves
guest version a major version to ensure that property is turned on.

There is a further solution to take a snapshot of the .config for the
primary pool. This would allow for the network association to be more
likely recreated correctly. However this would require either the
entire chyves dataset snapped recursively at the same time if using
‘zfs snapshot’ manually or chyves snapshotting that dataset when taking
one of a guest. Doable, but not too appealing.