FIX: Moved network adjustment to end of snapshot rollback function.
Ran into issue with 88 character limit with folder path. Basically the
snapshot state configuration could not pull but it would run anyways.
This would cause all tap interfaces to be unjoined from bridge
associations.

Moved all network adjustment functionality to after the rollback. Not
preferred as I would rather it be ran before and after to maximize
success. However a comprehensive workaround was not thought of in two
days, this will have to do for now.