ENH: New property 'eject_iso_on_n_reboot'. DATASET+
This ejects an ISO from a guest after {n} reboots.

This is both a global and guest property. When the property is set on
the guest it overrides the global property. The default the value is
set to ‘1’. The ‘chyves <guest> unset eject_iso_on_n_reboot’ removes
the property from the guest if desired.