FIX: Setting COMPILER_TYPE on FreeNAS required for make install
An error message is received when attempting to run ‘make install’ with
chyves.

[root@freenas] ~/chyves-dev# make install
make: "/usr/share/mk/bsd.compiler.mk" line 39: Unable to determine
compiler type for cc.  Consider setting COMPILER_TYPE.

As this is arbitrary for chyves, I do not see any harm in setting this
variable. May break in the future in ports for dependents.