FIX: Various fixes in 'chyves info'
At line 368 and 369, issue where I never updated string when breaking
out the size, cpu, and ram variables to show with the -b flag. This
never demonstrated as a problem as delimited characters must have a
character in-between them otherwise they are treated as a single column.

Updated and added code comments.

Fixed issue where disks would only show when the -z or -a flag were
used.