Moved clone handling in __parse_cmd_ingress
In the future I am looking to add the ability to clone a guest from a
snapshot state. This is the start of feature.

Cloning a guest is easy for the disks but harder for the guest
properties. This may run into the 88 character limit again.

Added ‘__verify_guests $1’ to the handling, otherwise unchanged.