Version increment to v0.1.6-dev for recent changes.
My GitHub desktop application synced changes early and before this
commit was made. This could have causes a discrepancy for users who
cloned or downloaded the repo archive from 9/3/2016 to 9/5/2016. My
apologies, it is my intention to always push any changes with a new
version number attached to it for debug reasons.