FIX: Be mindful of MTU for network bridge.
This now adds the physical interface to the bridge first. This sets the
MTU for the bridge per the man page for if_bridge(4), “The MTU of the
first member interface to be added is used as the bridge MTU.”.

Then when a device is added to a bridge, the MTU is set based on the
MTU of the bridge when not 1500.

This is to address:
https://github.com/chyves/chyves/issues/7