ENH: Added VyOS support for grub-bhyve
Will commit the doc portion in the next commit as there is a change to
the OpenBSD handling. Major CGV increment for OpenBSD changes.

Alphabetically reordered grep string for user input verification for
‘os’ property.

When VyOS is installed to disk, VyOS does not need any special
grub-bhyve handling configuration and shares the ‘default’ values.